separate out inference from ophys_processing dag since requires a trained model, and the trained model requires ophys_processing outputs to train
